Move jobs and wrapper interfaces into app layer
Jobs are injected into so they are incorrect in living in the dagger layer
