+++ src/main/java/com/pyamsoft/powermanager/app/job/DozeManageJob.java
@@ -14,23 +14,22 @@
  * limitations under the License.
  */
 
-package com.pyamsoft.powermanager.dagger.job;
+package com.pyamsoft.powermanager.app.job;
 
 import android.support.annotation.CheckResult;
 import android.support.annotation.NonNull;
 import com.birbit.android.jobqueue.Job;
 import com.pyamsoft.powermanager.PowerManagerSingleInitProvider;
 import com.pyamsoft.powermanager.app.modifier.BooleanInterestModifier;
-import com.pyamsoft.powermanager.app.observer.BooleanInterestObserver;
-import com.pyamsoft.powermanager.dagger.wrapper.JobSchedulerCompat;
+import com.pyamsoft.powermanager.app.wrapper.JobSchedulerCompat;
 import javax.inject.Inject;
 import javax.inject.Named;
 
-public abstract class BluetoothManageJob extends ManageJob {
+public abstract class DozeManageJob extends ManageJob {
 
-  @NonNull public static final String JOB_TAG = "bluetooth_job";
+  @NonNull public static final String JOB_TAG = "doze_job";
 
-  @SuppressWarnings("WeakerAccess") BluetoothManageJob(@NonNull JobSchedulerCompat jobManager,
+  @SuppressWarnings("WeakerAccess") DozeManageJob(@NonNull JobSchedulerCompat jobManager,
       @NonNull JobType jobType, long delayInSeconds, boolean periodic, long periodicEnableInSeconds,
       long periodicDisableInSeconds) {
     super(jobManager, JOB_TAG, jobType, delayInSeconds, periodic, periodicEnableInSeconds,
@@ -49,10 +48,9 @@ public abstract class BluetoothManageJob extends ManageJob {
         periodicEnableInSeconds, periodicDisableInSeconds);
   }
 
-  public static final class EnableJob extends BluetoothManageJob {
+  public static final class EnableJob extends DozeManageJob {
 
-    @Inject @Named("mod_bluetooth_state") BooleanInterestModifier interestModifier;
-    @Inject @Named("obs_bluetooth_state") BooleanInterestObserver interestObserver;
+    @Inject @Named("mod_doze_state") BooleanInterestModifier interestModifier;
 
     EnableJob(@NonNull JobSchedulerCompat jobManager, long delayTimeInMillis, boolean periodic,
         long periodicEnableInSeconds, long periodicDisableInSeconds) {
@@ -71,16 +69,15 @@ public abstract class BluetoothManageJob extends ManageJob {
     }
 
     @Override public void run() {
-      if (!interestObserver.is()) {
-        interestModifier.set();
-      }
+      // Doze job is a bit backwards since Doze is thought of differently
+      // Doze being 'enabled' actually means to turn it off
+      interestModifier.unset();
     }
   }
 
-  public static final class DisableJob extends BluetoothManageJob {
+  public static final class DisableJob extends DozeManageJob {
 
-    @Inject @Named("mod_bluetooth_state") BooleanInterestModifier interestModifier;
-    @Inject @Named("obs_bluetooth_state") BooleanInterestObserver interestObserver;
+    @Inject @Named("mod_doze_state") BooleanInterestModifier interestModifier;
 
     DisableJob(@NonNull JobSchedulerCompat jobManager, long delayTimeInMillis, boolean periodic,
         long periodicEnableInSeconds, long periodicDisableInSeconds) {
@@ -102,9 +99,9 @@ public abstract class BluetoothManageJob extends ManageJob {
     }
 
     @Override public void run() {
-      if (interestObserver.is()) {
-        interestModifier.unset();
-      }
+      // Doze job is a bit backwards since Doze is thought of differently
+      // Doze being 'disabled' actually means to turn it on
+      interestModifier.set();
     }
   }
 }
