+++ src/main/java/com/pyamsoft/powermanager/app/job/SyncManageJob.java
@@ -14,25 +14,26 @@
  * limitations under the License.
  */
 
-package com.pyamsoft.powermanager.dagger.job;
+package com.pyamsoft.powermanager.app.job;
 
 import android.support.annotation.CheckResult;
 import android.support.annotation.NonNull;
 import com.birbit.android.jobqueue.Job;
 import com.pyamsoft.powermanager.PowerManagerSingleInitProvider;
 import com.pyamsoft.powermanager.app.modifier.BooleanInterestModifier;
-import com.pyamsoft.powermanager.dagger.wrapper.JobSchedulerCompat;
+import com.pyamsoft.powermanager.app.observer.BooleanInterestObserver;
+import com.pyamsoft.powermanager.app.wrapper.JobSchedulerCompat;
 import javax.inject.Inject;
 import javax.inject.Named;
 
-public abstract class DozeManageJob extends ManageJob {
+public abstract class SyncManageJob extends ManageJob {
 
-  @NonNull public static final String JOB_TAG = "doze_job";
+  @NonNull public static final String JOB_TAG = "sync_job";
 
-  @SuppressWarnings("WeakerAccess") DozeManageJob(@NonNull JobSchedulerCompat jobManager,
+  @SuppressWarnings("WeakerAccess") SyncManageJob(@NonNull JobSchedulerCompat jobSchedulerCompat,
       @NonNull JobType jobType, long delayInSeconds, boolean periodic, long periodicEnableInSeconds,
       long periodicDisableInSeconds) {
-    super(jobManager, JOB_TAG, jobType, delayInSeconds, periodic, periodicEnableInSeconds,
+    super(jobSchedulerCompat, JOB_TAG, jobType, delayInSeconds, periodic, periodicEnableInSeconds,
         periodicDisableInSeconds);
   }
 
@@ -48,19 +49,20 @@ public abstract class DozeManageJob extends ManageJob {
         periodicEnableInSeconds, periodicDisableInSeconds);
   }
 
-  public static final class EnableJob extends DozeManageJob {
+  public static final class EnableJob extends SyncManageJob {
 
-    @Inject @Named("mod_doze_state") BooleanInterestModifier interestModifier;
+    @Inject @Named("mod_sync_state") BooleanInterestModifier interestModifier;
+    @Inject @Named("obs_sync_state") BooleanInterestObserver interestObserver;
 
-    EnableJob(@NonNull JobSchedulerCompat jobManager, long delayTimeInMillis, boolean periodic,
-        long periodicEnableInSeconds, long periodicDisableInSeconds) {
-      super(jobManager, JobType.ENABLE, delayTimeInMillis, periodic, periodicEnableInSeconds,
-          periodicDisableInSeconds);
+    EnableJob(@NonNull JobSchedulerCompat jobSchedulerCompat, long delayTimeInMillis,
+        boolean periodic, long periodicEnableInSeconds, long periodicDisableInSeconds) {
+      super(jobSchedulerCompat, JobType.ENABLE, delayTimeInMillis, periodic,
+          periodicEnableInSeconds, periodicDisableInSeconds);
     }
 
     @CheckResult @NonNull
-    public static EnableJob createManagerEnableJob(@NonNull JobSchedulerCompat jobManager) {
-      return new EnableJob(jobManager, 100L, false, 0, 0);
+    public static EnableJob createManagerEnableJob(@NonNull JobSchedulerCompat jobSchedulerCompat) {
+      return new EnableJob(jobSchedulerCompat, 100L, false, 0, 0);
     }
 
     @Override public void onAdded() {
@@ -69,28 +71,29 @@ public abstract class DozeManageJob extends ManageJob {
     }
 
     @Override public void run() {
-      // Doze job is a bit backwards since Doze is thought of differently
-      // Doze being 'enabled' actually means to turn it off
-      interestModifier.unset();
+      if (!interestObserver.is()) {
+        interestModifier.set();
+      }
     }
   }
 
-  public static final class DisableJob extends DozeManageJob {
+  public static final class DisableJob extends SyncManageJob {
 
-    @Inject @Named("mod_doze_state") BooleanInterestModifier interestModifier;
+    @Inject @Named("mod_sync_state") BooleanInterestModifier interestModifier;
+    @Inject @Named("obs_sync_state") BooleanInterestObserver interestObserver;
 
-    DisableJob(@NonNull JobSchedulerCompat jobManager, long delayTimeInMillis, boolean periodic,
-        long periodicEnableInSeconds, long periodicDisableInSeconds) {
-      super(jobManager, JobType.DISABLE, delayTimeInMillis, periodic, periodicEnableInSeconds,
-          periodicDisableInSeconds);
+    DisableJob(@NonNull JobSchedulerCompat jobSchedulerCompat, long delayTimeInMillis,
+        boolean periodic, long periodicEnableInSeconds, long periodicDisableInSeconds) {
+      super(jobSchedulerCompat, JobType.DISABLE, delayTimeInMillis, periodic,
+          periodicEnableInSeconds, periodicDisableInSeconds);
     }
 
     @CheckResult @NonNull
-    public static DisableJob createManagerDisableJob(@NonNull JobSchedulerCompat jobManager,
+    public static DisableJob createManagerDisableJob(@NonNull JobSchedulerCompat jobSchedulerCompat,
         long delayTimeInMillis, boolean periodic, long periodicEnableInSeconds,
         long periodicDisableInSeconds) {
-      return new DisableJob(jobManager, delayTimeInMillis, periodic, periodicEnableInSeconds,
-          periodicDisableInSeconds);
+      return new DisableJob(jobSchedulerCompat, delayTimeInMillis, periodic,
+          periodicEnableInSeconds, periodicDisableInSeconds);
     }
 
     @Override public void onAdded() {
@@ -99,9 +102,9 @@ public abstract class DozeManageJob extends ManageJob {
     }
 
     @Override public void run() {
-      // Doze job is a bit backwards since Doze is thought of differently
-      // Doze being 'disabled' actually means to turn it on
-      interestModifier.set();
+      if (interestObserver.is()) {
+        interestModifier.unset();
+      }
     }
   }
 }
