+++ src/main/java/com/pyamsoft/powermanager/app/service/job/PowerManagerGCMJobSchedulerService.java
@@ -20,7 +20,7 @@ import android.support.annotation.NonNull;
 import com.birbit.android.jobqueue.JobManager;
 import com.birbit.android.jobqueue.scheduling.GcmJobSchedulerService;
 import com.pyamsoft.powermanager.PowerManagerSingleInitProvider;
-import com.pyamsoft.powermanager.dagger.wrapper.JobSchedulerCompat;
+import com.pyamsoft.powermanager.app.wrapper.JobSchedulerCompat;
 import javax.inject.Inject;
 
 public class PowerManagerGCMJobSchedulerService extends GcmJobSchedulerService {
