+++ src/main/java/com/pyamsoft/powermanager/dagger/manager/WearAwareManagerBaseInteractor.java
@@ -19,7 +19,7 @@ package com.pyamsoft.powermanager.dagger.manager;
 import android.support.annotation.NonNull;
 import com.pyamsoft.powermanager.PowerManagerPreferences;
 import com.pyamsoft.powermanager.app.observer.BooleanInterestObserver;
-import com.pyamsoft.powermanager.dagger.wrapper.JobSchedulerCompat;
+import com.pyamsoft.powermanager.app.wrapper.JobSchedulerCompat;
 import rx.Observable;
 import timber.log.Timber;
