+++ src/main/java/com/pyamsoft/powermanager/dagger/modifier/state/BluetoothStateModifier.java
@@ -19,7 +19,7 @@ package com.pyamsoft.powermanager.dagger.modifier.state;
 import android.content.Context;
 import android.support.annotation.NonNull;
 import com.pyamsoft.powermanager.PowerManagerPreferences;
-import com.pyamsoft.powermanager.dagger.wrapper.DeviceFunctionWrapper;
+import com.pyamsoft.powermanager.app.wrapper.DeviceFunctionWrapper;
 import javax.inject.Inject;
 
 class BluetoothStateModifier extends StateModifier {
