+++ src/main/java/com/pyamsoft/powermanager/dagger/wrapper/BluetoothAdapterWrapperImpl.java
@@ -23,6 +23,7 @@ import android.os.Build;
 import android.support.annotation.CheckResult;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
+import com.pyamsoft.powermanager.app.wrapper.DeviceFunctionWrapper;
 import javax.inject.Inject;
 import timber.log.Timber;
