+++ src/main/java/com/pyamsoft/powermanager/dagger/wrapper/JobSchedulerCompatModule.java
@@ -18,6 +18,7 @@ package com.pyamsoft.powermanager.dagger.wrapper;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
+import com.pyamsoft.powermanager.app.wrapper.JobSchedulerCompat;
 import dagger.Module;
 import dagger.Provides;
 import javax.inject.Singleton;
