+++ src/main/java/com/pyamsoft/powermanager/dagger/wrapper/WifiManagerWrapperImpl.java
@@ -21,6 +21,7 @@ import android.net.wifi.WifiManager;
 import android.support.annotation.CheckResult;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
+import com.pyamsoft.powermanager.app.wrapper.DeviceFunctionWrapper;
 import javax.inject.Inject;
 import timber.log.Timber;
