+++ src/main/java/com/pyamsoft/powermanager/dagger/wrapper/WrapperModule.java
@@ -18,6 +18,7 @@ package com.pyamsoft.powermanager.dagger.wrapper;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
+import com.pyamsoft.powermanager.app.wrapper.DeviceFunctionWrapper;
 import dagger.Module;
 import dagger.Provides;
 import javax.inject.Named;
